Schema changes: Added a new RatioQuantLayer with the DataType element.
Added a new optional rawFile_ref to Feature.

ACTION - Da, can you please check if the use of xsd:unique and xsd:keyref etc needs to be updated for RatioQLType - I was forced to a new identifier for the xsd:unique type within RQLType but I don't understand the meaning of this.


Former-commit-id: 6f8608bbf4c9a86fd16c0779e43daa80edbbaccc